# services/mr-elephant/docs/swagger.yaml
basePath: /v1
definitions:
  github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.HistoryDTO:
    properties:
      history_id:
        type: string
      id:
        type: integer
      success:
        type: boolean
      task_id:
        type: string
      task_options:
        type: string
      task_result:
        type: string
      task_type:
        type: string
    type: object
  github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.ResultDTO:
    properties:
      command:
        type: string
      contents:
        type: string
      id:
        type: integer
      result_id:
        type: string
      success:
        type: boolean
      task_id:
        type: string
    type: object
  github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.TaskDTO:
    properties:
      command:
        type: string
      id:
        type: integer
      task_id:
        type: string
      task_type:
        type: string
    type: object
host: http://127.0.0.1:4000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a server R2 server listener.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Mr Elephant API
  version: "1.0"
paths:
  /history:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all history entries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.HistoryDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.ErrorResponse'
      summary: Get all history
      tags:
      - history
  /result:
    post:
      consumes:
      - application/json
      description: Create a new result with the given details
      parameters:
      - description: Result to create
        in: body
        name: result
        required: true
        schema:
          $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.ResultDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.ResultDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.ErrorResponse'
      summary: Create a new result
      tags:
      - results
  /results:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all results
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.ResultDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.ErrorResponse'
      summary: Get all results
      tags:
      - results
  /task:
    post:
      consumes:
      - application/json
      description: Create a new task with the given details
      parameters:
      - description: Task to create
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.TaskDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.TaskDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.ErrorResponse'
      summary: Create a new task
      tags:
      - tasks
  /tasks:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.TaskDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mauricetjmurphy_mr-elephant_internal_api_v1_models.ErrorResponse'
      summary: Get all tasks
      tags:
      - tasks
swagger: "2.0"